Moved Dataset to base in test
To prevent duplication of code, move dataset to base and import for each
model.
